src/backend/requirements.txt
django
djangorestframework
django-rest-enumfield
drf-spectacular
APScheduler
django-cors-headers
uwsgi